feat(refactor): refactor and organize project
- create a handlers folder and separate the logic for handling the user and screams routes
- create a utils folder and a separate file for handling the authentication middleware
- create a validators file and put th input validation logic in it
- create a config file and put the firebase config object in it
