+++ functions/util/config.js
@@ -1,0 +1,9 @@
+module.exports = {
+  apiKey: "AIzaSyDqglVH8NEhhBUeIoI1urDNS52B_fhTp0g",
+  authDomain: "amistad-9f94a.firebaseapp.com",
+  databaseURL: "https://amistad-9f94a.firebaseio.com",
+  projectId: "amistad-9f94a",
+  storageBucket: "amistad-9f94a.appspot.com",
+  messagingSenderId: "1088613517092",
+  appId: "1:1088613517092:web:763fda8d6a29fa82"
+};
